+++ database/migrations/2025_03_17_142200_add_maya_subscription_fields_to_user_subscriptions_table.php
@@ -3,7 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
-
+use Illuminate\Support\Facades\DB;
 return new class extends Migration
 {
     /**
@@ -24,13 +24,18 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('user_subscriptions', function (Blueprint $table) {
-            $table->dropColumn([
-                'maya_subscription_id',
-                'maya_payment_token_id',
-                'next_billing_date',
-                'billing_cycle',
-            ]);
-        });
+        if (Schema::hasTable('user_subscriptions')) {
+            $count = DB::table('user_subscriptions')->count();
+            if ($count === 0) {
+                Schema::table('user_subscriptions', function (Blueprint $table) {
+                    $table->dropColumn([
+                        'maya_subscription_id',
+                        'maya_payment_token_id',
+                        'next_billing_date',
+                        'billing_cycle',
+                    ]);
+                });
+            }
+        }
     }
 };
